+++ CHANGELOG.md
@@ -4,10 +4,9 @@
 
 ### Fixed
 
-- Fix regression in Scrollspy (IE11)
+- Fix regression in Scrollspy in IE 11
 - Fix Sass distribution
 - Fix large close icon
-
 
 ## 3.0.0 rc 1 (May 15, 2018)